ittzm

+++ lab12_1.cpp
@@ -1,7 +1,7 @@
 #include<iostream>
 using namespace std;
 
-int fibonacci(int);
+long long int fibonacci(int);
 
 int main(){
     int x;
@@ -10,7 +10,7 @@ int main(){
     return 0; 
 }
 
-int fibonacci(int x){
+long long int fibonacci(int x){
     if(x==0) return 0;
     else if(x==1) return 1;
     else{
